+++ NEWS.md
@@ -1,6 +1,15 @@
-# sircovid 0.10.21
+# sircovid 0.10.24
 
 * New set of parameter `strain_rel_gamma_` for varying `gamma_A`, `gamma_P`, `gamma_C_1`, and `gamma_C_2` by strain.
+
+# sircovid 0.10.23
+
+* Support to allow catch up of vaccine doses that could not be delivered 
+(e.g. when people were infected at the time they should have been vaccinated)
+
+# sircovid 0.10.22
+
+* Added support for time-varying hospital durations
 
 # sircovid 0.10.20